# docs/swagger.yaml
definitions:
  models.Address:
    properties:
      country:
        type: string
      geolocation:
        $ref: '#/definitions/models.MongoGeolocation'
      houseNumber:
        type: string
      locality:
        type: string
      postCode:
        type: string
      state:
        type: string
      street:
        type: string
    type: object
  models.Event:
    properties:
      address:
        $ref: '#/definitions/models.Address'
      city:
        example: SuperCity
        type: string
      comment:
        example: Super exciting comment.
        type: string
      country:
        example: SuperCountry
        type: string
      date:
        example: "2021-10-31T19:00:00.000Z"
        type: string
      genres:
        example:
        - german trap
        items:
          type: string
        type: array
      genresText:
        example: begleitet von diversen Berner Hip-Hop Acts. Von Trap und Phonk bis
          zu Afrobeats - Free Quenzy's Produktionen bieten eine breite Palette an
          Sounds.
        type: string
      imageUrl:
        example: http://link.to/concert/image.jpg
        type: string
      location:
        example: SuperLocation
        type: string
      offset:
        type: integer
      sourceUrl:
        example: http://link.to/source
        type: string
      title:
        example: ExcitingTitle
        type: string
      type:
        example: concert
        type: string
      url:
        example: http://link.to/concert/page
        type: string
    required:
    - city
    - date
    - location
    - sourceUrl
    - title
    - type
    - url
    type: object
  models.MongoGeolocation:
    properties:
      coordinates:
        items:
          type: number
        type: array
      type:
        type: string
    type: object
  models.ScraperStatus:
    properties:
      lastScrapeEnd:
        example: "2021-10-31T19:05:00.000Z"
        type: string
      lastScrapeStart:
        example: "2021-10-31T19:00:00.000Z"
        type: string
      nrErrors:
        example: 0
        type: integer
      nrItems:
        example: 100
        type: integer
      scraperLogs:
        type: string
      scraperName:
        example: Helsinki
        type: string
    required:
    - lastScrapeEnd
    - lastScrapeStart
    - nrErrors
    - nrItems
    - scraperName
    type: object
info:
  contact: {}
paths:
  /api/events:
    delete:
      consumes:
      - application/json
      description: Delete events.
      parameters:
      - description: sourceUrl string
        in: query
        name: sourceUrl
        type: string
      - description: datetime string
        in: query
        name: datetime
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: A success message
          schema:
            type: string
        "500":
          description: failed to delete events
          schema:
            type: string
      security:
      - BasicAuth: []
      summary: Delete events.
      tags:
      - events
    get:
      consumes:
      - application/json
      description: This endpoint returns all events matching the search terms. Note
        that only events from today on will be returned if no date is passed, ie no
        past events.
      parameters:
      - description: title search string
        in: query
        name: title
        type: string
      - description: location search string
        in: query
        name: location
        type: string
      - description: type search string
        in: query
        name: type
        type: string
      - description: city search string
        in: query
        name: city
        type: string
      - description: country search string
        in: query
        name: country
        type: string
      - description: radius around given city in kilometers
        in: query
        name: radius
        type: integer
      - description: date search string
        in: query
        name: date
        type: string
      - description: page number
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Event'
            type: array
        "404":
          description: No events found
          schema:
            type: string
      summary: Get all events.
      tags:
      - events
    post:
      consumes:
      - application/json
      description: Add new events to the database.
      parameters:
      - description: Event Info
        in: body
        name: message
        required: true
        schema:
          items:
            $ref: '#/definitions/models.Event'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: A json with the results
          schema:
            type: string
        "400":
          description: failed to parse body
          schema:
            type: string
        "500":
          description: failed to insert events
          schema:
            type: string
      security:
      - BasicAuth: []
      summary: Add new events.
      tags:
      - events
  /api/events/{field}:
    get:
      description: This endpoint returns all distinct values for the given field.
        Note that past events are not considered for this query.
      parameters:
      - description: field name, can only be location or city
        in: path
        name: field
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: failed to retrieve values
          schema:
            type: string
      summary: Get distinct field values.
      tags:
      - events
  /api/events/today/slack:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: This endpoint returns today's events for a given city in a format
        that slack needs for its slash command.
      parameters:
      - in: formData
        name: text
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: A json with the results
          schema:
            type: string
      summary: Get today's events.
      tags:
      - events
  /api/events/validate:
    post:
      consumes:
      - application/json
      description: This endpoint validates events.
      parameters:
      - description: Event Info
        in: body
        name: message
        required: true
        schema:
          items:
            $ref: '#/definitions/models.Event'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: A json with the results
          schema:
            type: string
        "400":
          description: failed to validate events
          schema:
            type: string
      summary: Validate events.
      tags:
      - events
  /api/notifications/activate:
    get:
      description: This endpoint activates a notification that has been added previously
        if the inactive notification hasn't expired yet (expires after 24h).
      parameters:
      - description: email
        in: query
        name: email
        type: string
      - description: token
        in: query
        name: token
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: failed to activate notification
          schema:
            type: string
        "500":
          description: failed to activate notification
          schema:
            type: string
      summary: Activate notification.
      tags:
      - notifications
  /api/notifications/add:
    get:
      description: Add new notification to the database.
      parameters:
      - description: title search string
        in: query
        name: title
        type: string
      - description: location search string
        in: query
        name: location
        type: string
      - description: city search string
        in: query
        name: city
        type: string
      - description: country search string
        in: query
        name: country
        type: string
      - description: radius around given city in kilometers
        in: query
        name: radius
        type: integer
      - description: email
        in: query
        name: email
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Failed to parse body
          schema:
            type: string
        "500":
          description: Failed to insert notification
          schema:
            type: string
      summary: Add new notification.
      tags:
      - notifications
  /api/notifications/delete:
    get:
      description: This endpoint deletes a notification that has been added previously
        based on the email address and the token.
      parameters:
      - description: email
        in: query
        name: email
        type: string
      - description: token
        in: query
        name: token
        type: string
      produces:
      - application/json
      responses:
        "500":
          description: Failed to delete notification
          schema:
            type: string
      summary: Delete notification.
      tags:
      - notifications
  /api/notifications/deleteInactive:
    delete:
      description: This endpoint deletes all inactive notification that are older
        than 24h.
      produces:
      - application/json
      responses:
        "500":
          description: Failed to delete notifications
          schema:
            type: string
      security:
      - BasicAuth: []
      summary: Delete inactive notifications.
      tags:
      - notifications
  /api/notifications/send:
    get:
      description: This endpoint sends an email for every active notification whose
        query returns a result.
      produces:
      - application/json
      responses:
        "500":
          description: failed to send notifications
          schema:
            type: string
      security:
      - BasicAuth: []
      summary: Send notifications.
      tags:
      - notifications
  /api/status:
    get:
      consumes:
      - application/json
      description: This endpoint returns all scraper statuses matching the search
        terms.
      parameters:
      - description: scraper name search string
        in: query
        name: name
        type: string
      - description: page number
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ScraperStatus'
            type: array
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: No scraper status found
          schema:
            type: string
      summary: Get scraper status.
      tags:
      - scraper status
    post:
      consumes:
      - application/json
      description: This endpoint inserts or updates the status of a scraper, based
        on the given name. For now, per scraper only one status is allowed.
      parameters:
      - description: Scraper status object
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/models.ScraperStatus'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ScraperStatus'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BasicAuth: []
      summary: Update or insert scraper status.
      tags:
      - scraper status
  /api/status/{name}:
    delete:
      description: This endpoint deletes the status of a scraper by name.
      parameters:
      - description: Scraper name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Status deleted successfully
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: Status not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BasicAuth: []
      summary: Delete scraper status.
      tags:
      - scraper status
swagger: "2.0"